refactor: update import paths for project, contributor, and contribution cards for consistency
