+++ web/src/components/contributor-card.tsx
