+++ web/src/pages/team/index.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { useEffect } from "react";
 import { Helmet } from "react-helmet-async";
+import ContributorCard from "src/components/contributor-card";
 import { Loading } from "src/components/loading";
 import { Locale, useLocale } from "src/components/locale";
 import { TryAgain } from "src/components/try-again";
 import { fetchContributorsListAction } from "src/redux/actions/contributors";
 import { useAppDispatch, useAppSelector } from "src/redux/store";
-import ContributorCard from "./contributor-card";
 
 // ts-prune-ignore-next
 export default function Page(): JSX.Element {
